Wednesday Second commit
To do list:

Limit the number of cards that can be drawn to 10 to match the number of cards that can be played. (DONE 12:30pm)

Add the ability to click on the cards in the dropzone to flip them. (DONE 1:20pm)

Restrict players from flipping cards they do not own. (DONE 1:35PM)

Need to fix card flipping on both server and client side